+++ src/pages/fruitsShop.svelte
